mds_data_sync/mds2mongo/requirements.txt
deepdiff
pymongo
python-dotenv
requests
